requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-pinecone
langchain-huggingface
langchain-community
pydantic
python-dotenv
starlette
supabase
pinecone-client
sentence-transformers
lxml # for HTML parsing
itsdangerous


